tests/wpt/tests/html/semantics/scripting-1/the-script-element/text-module/referrer-policies.sub.html

<!DOCTYPE html>
<html>
<head>
<title>Referrers with text module requests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script type="module">
  // "name" parameter is necessary for bypassing the module map.
  import referrerSame from "./referrer-checker.py?name=sameNoReferrerPolicy" with { type: "text" };
  import referrerRemote from "http://{{domains[www1]}}:{{ports[http][0]}}/html/semantics/scripting-1/the-script-element/text-module/referrer-checker.py?name=remoteNoReferrerPolicy" with { type: "text" };

  const origin = (new URL(location.href)).origin + "/";
  const originUrl = location.href;

  test(t => {
    assert_equals(
        referrerSame, originUrl,
        "Referrer URL should be sent for the same-origin top-level script.");
  }, "Importing a same-origin top-level script with the default referrer policy.");

  test(t => {
    assert_equals(
        referrerRemote, origin,
        "Referrer origin should be sent for the remote-origin top-level script.");
  }, "Importing a remote-origin top-level script with the default referrer policy.");
</script>
<script type="module" referrerpolicy="origin">
  import referrerSame from "./referrer-checker.py?name=sameReferrerPolicyOrigin" with { type: "text" };
  import referrerRemote from "http://{{domains[www1]}}:{{ports[http][0]}}/html/semantics/scripting-1/the-script-element/text-module/referrer-checker.py?name=remoteReferrerPolicyOrigin" with { type: "text" };

  const origin = (new URL(location.href)).origin + "/";

  test(t => {
    assert_equals(
        referrerSame, origin,
        "Referrer origin should be sent for the same-origin top-level script.");
  }, "Importing a same-origin top-level script with the origin policy.");

  test(t => {
    assert_equals(
        referrerRemote, origin,
        "Referrer origin should be sent for the remote-origin top-level script.");
  }, "Importing a remote-origin top-level script with the origin policy.");

</script>
<script type="module" referrerpolicy="no-referrer">
  import referrerSame from "./referrer-checker.py?name=sameReferrerPolicyNoReferrer" with { type: "text" };
  import referrerRemote from "http://{{domains[www1]}}:{{ports[http][0]}}/html/semantics/scripting-1/the-script-element/text-module/referrer-checker.py?name=remoteReferrerPolicyNoReferrer" with { type: "text" };

  test(t => {
    assert_equals(
        referrerSame, "",
        "No referrer should be sent for the same-origin top-level script.");
  }, "Importing a same-origin top-level script with the no-referrer policy.");

  test(t => {
    assert_equals(
        referrerRemote, "",
        "No referrer should be sent for the remote-origin top-level script.");
  }, "Importing a remote-origin top-level script with the no-referrer policy.");

</script>
<script type="module" referrerpolicy="unsafe-url">
  import referrerSame from "./referrer-checker.py?name=sameNoReferrerPolicyUnsafeUrl" with { type: "text" };
  import referrerRemote from "http://{{domains[www1]}}:{{ports[http][0]}}/html/semantics/scripting-1/the-script-element/text-module/referrer-checker.py?name=remoteNoReferrerPolicyUnsafeUrl" with { type: "text" };

  const originUrl = location.href;

  test(t => {
    assert_equals(
        referrerSame, originUrl,
        "Referrer URL should be sent for the same-origin top-level script.");
  }, "Importing a same-origin top-level script with the unsafe-url referrer policy.");

  test(t => {
    assert_equals(
        referrerRemote, originUrl,
        "Referrer URL should be sent for the remote-origin top-level script.");
  }, "Importing a remote-origin top-level script with the unsafe-url referrer policy.");
</script>
</body>
</html>
